backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
pydantic-settings==2.5.2
google-auth==2.35.0
requests==2.32.3
